% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phylowidget}
\alias{phylowidget}
\title{phylowidget Phylogeny widget}
\usage{
phylowidget(nwk, width = NULL, height = NULL)
}
\arguments{
\item{nwk}{Newick string or \code{phylo} from the \code{\link[ape]{ape}} package.}

\item{width}{Integer in px to override the default width. The default width fills the browser window.}

\item{height}{Integer in px to override the default height. The default height fills the browser window.}
}
\description{
An interactive phylogeny viewer using D3.js, based on phylotree.js. Just supply
a \code{Newick} string or \code{phylo} data as the parameter.
}
\note{
This displays a phylogeny in a browser window; by default, the viewer is disabled.
}
\examples{
\dontrun{
# A stand-alone example
library(ape)
data(bird.orders)
phylowidget(bird.orders)
}
}
\references{
\code{phylotree.js} \url{http://github.com/veg/phylotree.js}
}

