% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phylowidgetOutput}
\alias{phylowidgetOutput}
\title{phylowidgetOutput Helper function for phyloshiny}
\usage{
phylowidgetOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output identifier.}

\item{width}{The width of the device.}

\item{height}{The height of the device.}
}
\description{
phylowidgetOutput Helper function for phyloshiny
}
\seealso{
renderPhylowidget
}

