% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylowidget.R
\name{phyloshiny}
\alias{phyloshiny}
\title{phyloshiny A phylowidget as a Shiny app}
\usage{
phyloshiny(nwk)
}
\arguments{
\item{nwk}{Either a Newick string or a \code{phylo} or \code{multiPhylo} object.}
}
\description{
phyloshiny A phylowidget as a Shiny app
}
\examples{
\dontrun{
library(ape)
data(bird.orders)
phyloshiny(bird.orders)
}
}
\seealso{
phylowidget
}

