% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylowidget-package.R
\docType{package}
\name{phylowidget-package}
\alias{phylowidget-package}
\title{phylowidget: htmlwidget wrapper for \href{https://github.com/veg/phylotree.js}{phylotree.js}}
\description{
phylowidget offers an interactive view of a phylogeny directly from R by
wrapping the Javascript library \href{https://github.com/veg/phylotree.js}{phylotree.js}
by Sergei Kosakovsky Pond and Steven Weaver. This combines the interactivity
of D3.js with the convenience and flexibility of an R \href{http://htmlwidgets.org}{htmlwidget}.
}
\details{
To learn more about phylowidget, start with the vignettes:
\code{browseVignettes(package = "phylowidget")}
}

