% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylowidget.R
\name{phylowidget}
\alias{phylowidget}
\title{phylowidget Phylogeny widget}
\usage{
phylowidget(nwk, width = NULL, height = NULL)
}
\arguments{
\item{nwk}{Either a Newick string or a \code{phylo} or \code{multiPhylo} object.}

\item{width}{The width of the device.}

\item{height}{The height of the device.}
}
\description{
An interactive phylogeny viewer using D3.js, based on phylotree.js
}
\note{
This displays a phylogeny in a browser window; by default, the viewer is disabled.
}
\examples{
## dontrun
# A stand-alone example
library(ape)
data(bird.orders)
phylowidget(bird.orders)
}
\references{
\code{phylotree.js} \url{http://github.com/veg/phylotree.js}
}

