% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylowidget.R
\name{renderPhylowidget}
\alias{renderPhylowidget}
\title{renderPhylowidget Helper function for phyloshiny}
\usage{
renderPhylowidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The function to be rendered}

\item{env}{The environment for rendering.}

\item{quoted}{Should the output be quoted?}
}
\description{
renderPhylowidget Helper function for phyloshiny
}
\seealso{
phylowidgetOutput
}

